Fix docstrings for the `wlm()` function
The return type annotation in the docstring describes `np.float64` values, but the actual code returns `float` values (line 466). The docstring should be updated to reflect the actual return type of `float` instead of `np.float64` to be consistent with the implementation.

+++ src/ts_stat_tests/algorithms/heteroscedasticity.py
@@ -419,12 +419,12 @@ def wlm(resid: ArrayLike, exog_het: ArrayLike) -> tuple[float, float, float, flo
             The explanatory variables for the variance. Usually, these are the original exogenous variables; the test internally handles adding their squares and cross-products.
 
     Returns:
-        (tuple[np.float64, np.float64, np.float64, np.float64]):
+        (tuple[float, float, float, float]):
             A tuple containing:
-            - `lmstat` (np.float64): The Lagrange Multiplier statistic.
-            - `lmpval` (np.float64): The p-value for the LM statistic.
-            - `fstat` (np.float64): The F-statistic.
-            - `fpval` (np.float64): The p-value for the F-statistic.
+            - `lmstat` (float): The Lagrange Multiplier statistic.
+            - `lmpval` (float): The p-value for the LM statistic.
+            - `fstat` (float): The F-statistic.
+            - `fpval` (float): The p-value for the F-statistic.
 
     ???+ example "Examples"
